uucore: Add time feature
We'll move some common time handling code there, to be shared
between du and ls as a start.

Starting with format_system_time from from `ls`.
