+++ src/uu/ls/Cargo.toml
@@ -24,11 +24,6 @@ ansi-width = { workspace = true }
 clap = { workspace = true, features = ["env"] }
 glob = { workspace = true }
 hostname = { workspace = true }
-jiff = { workspace = true, features = [
-  "tzdb-bundle-platform",
-  "tzdb-zoneinfo",
-  "tzdb-concatenated",
-] }
 lscolors = { workspace = true }
 selinux = { workspace = true, optional = true }
 terminal_size = { workspace = true }
@@ -41,6 +36,7 @@ uucore = { workspace = true, features = [
   "fsxattr",
   "parser",
   "quoting-style",
+  "time",
   "version-cmp",
 ] }
 uutils_term_grid = { workspace = true }
