+++ src/uucore/src/lib/features.rs
@@ -40,6 +40,8 @@ pub mod ranges;
 pub mod ringbuffer;
 #[cfg(feature = "sum")]
 pub mod sum;
+#[cfg(feature = "time")]
+pub mod time;
 #[cfg(feature = "update-control")]
 pub mod update_control;
 #[cfg(feature = "uptime")]
